Build system: correctly find libdivecomputer includes
Fully unsure when this got broken, but we tried to find the
libdivecomputer includes based on the include file "hw.h".
Interestingly, that file does not exisist (any more?) in
libdivecomputer, so the search for the include fails. This
is annoying, as the initial cmake fails on this in case of
developer builds from QtCreator (which do not compile all
dependencies like our home grown build scripts).

The solution is simple: just find the includes for
libdivecomputer based on exiting files in this lib.

Signed-off-by: Jan Mulder <jlmulder@xs4all.nl>