Dive pictures: Add videoThumbnailCache
If a video is identified, add a picture to videoThumbnailCache.
Currently, this is only a null QImage, which is used as a flag.
In the future, actually extract a still frame from videos.

In divepicturemodel, check for entries in videoThumbnailCache
and set the new isVideo flag in the PictureEntry struct.
Since no stills are extracted, use a dummy thumbnail instead.

Videos are identified according to a hard-coded list if filename
extensions. These extensions are added to the list of supported
extensions in the image-open dialog.

Signed-off-by: Berthold Stoeger <bstoeger@mail.tuwien.ac.at>