Dive pictures: play videos in profile plot
On mouse-over, play videos in the profile plot thumbnails. This is
realized by deriving from QAbstractVideoSurface and setting the
frames as images. On click, stop playing videos, to avoid playing
videos simultaneously in the profile and the system viewers.

Don't compile video playback for mobile, since dive pictures are
not yet supported anyway.

For a yet unknown reason, using QGraphicsVideoItem did not work on
Linux.

Signed-off-by: Berthold Stoeger <bstoeger@mail.tuwien.ac.at>