qthelper.cpp: pass copy of QString to hashFile()
This is a partial revert of e5dcd9fc161891a4e70364e1dcdf232590eb49c6,
which replaced a number of QString by-value arguments by
reference-to-QString arguments.

This introduced a bug in the case of hashFile, which is called from
a different thread context. The passed QString might be deleted in
the calling function, leading to a dangling reference.

Signed-off-by: Berthold Stoeger <bstoeger@mail.tuwien.ac.at>