Cleanup: delete stale DivePictureItems right away
The function ProfileWidget2::plotPictures() replots all pictures
in the profile plot. Before doing so, it deletes the previously
displayed items.

This was done using deleteLater(), which did not free the memory
immediately, but only later in the event loop. Change this to properly
delete the item.

By deleting the item right away, it is not necessary to hide the item
beforehand. Remove the corresponding function call.

Signed-off-by: Berthold Stoeger <bstoeger@mail.tuwien.ac.at>