Planner: calculate last manually entered sample on demand
Commit 9a8bab21c9996020b16f1dd140647efe8e262a42 introduced calculation
of the time of the last manually entered sample. The problem is
that the calculation was done in create_dive_from_plan(), but
samples could also be created in fake_dc(), which resulted in an
invalid last-manually-entered-sample time. This in return led to
the issue reported in #1211.

To fix this problem, and to remove the complexity of keeping track
of a cached last manually entered sample, calculate this datum on
demand. There seems to be no point in keeping track of this
redundant piece of information.

This improves the issue reported in #1211 significantly, but does
not fix it completely, because the conversion duration->samples->duration
is not a no-op.

Signed-off-by: Berthold Stoeger <bstoeger@mail.tuwien.ac.at>