Dive pictures: Derive thumbnail file from picture filename
Since commit 6618c9ebfc6a7cebbef687fcb3aa74c70f504ff2, thumbnails
are saved in individual files. The filename was simply the picture-hash.
In a mailing-list discussion it turned out that in the future we might
not hash images or change the hash. Therefore, derive the thumbnail
filename from the image filename, using the SHA1 algorithm.

Signed-off-by: Berthold Stoeger <bstoeger@mail.tuwien.ac.at>