Desktop: refactor DiveComputerModel
The divecomputer list is stored as a QMultiMap. Since Qt accesses
the table rows via indexes, on every call the QMultiMap was
transformed into a QList. Instead introduce a function that steps
through the QMultiMap.

Moreover, remove the DiveComputerModel::update(), which in principle
was only a very complicated way of setting the numRows member.
The numRows member is not necessary, since the number of rows can
be directly read from the QMultiMap.

Signed-off-by: Berthold Stoeger <bstoeger@mail.tuwien.ac.at>